requirements.txt
GitPython==2.1.1
toolz==0.10.0
termcolor==1.1.0
simplejson==3.17.0